infrastructure/streamlit/requirements.txt
streamlit
elasticsearch==8.14.0
psycopg2-binary==2.9.9
python-dotenv
openai==1.35.7
sentence-transformers==2.7.0
numpy==1.26.4
torch==2.3.1
scikit-learn 
pandas 
elasticsearch
transformers 
accelerate 
bitsandbytes 
sentencepiece
ollama